[refactor] #70 UserProblemResult를 submission 패키지로 이동
