+++ src/main/java/_ganzi/codoc/problem/service/ProblemService.java
@@ -5,19 +5,19 @@ import _ganzi.codoc.global.util.CursorPagingUtils;
 import _ganzi.codoc.problem.domain.Problem;
 import _ganzi.codoc.problem.domain.Quiz;
 import _ganzi.codoc.problem.domain.SummaryCard;
-import _ganzi.codoc.problem.domain.UserProblemResult;
 import _ganzi.codoc.problem.dto.ProblemListCondition;
 import _ganzi.codoc.problem.dto.ProblemListItem;
 import _ganzi.codoc.problem.dto.ProblemResponse;
 import _ganzi.codoc.problem.dto.ProblemSearchParam;
 import _ganzi.codoc.problem.enums.ProblemLevel;
-import _ganzi.codoc.problem.enums.ProblemSolvingStatus;
 import _ganzi.codoc.problem.exception.ProblemNotFoundException;
 import _ganzi.codoc.problem.repository.BookmarkRepository;
 import _ganzi.codoc.problem.repository.ProblemRepository;
 import _ganzi.codoc.problem.repository.QuizRepository;
 import _ganzi.codoc.problem.repository.SummaryCardRepository;
-import _ganzi.codoc.problem.repository.UserProblemResultRepository;
+import _ganzi.codoc.submission.domain.UserProblemResult;
+import _ganzi.codoc.submission.enums.ProblemSolvingStatus;
+import _ganzi.codoc.submission.repository.UserProblemResultRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
